Add files via upload
This is an educational game for guessing the states name of U.S. user has to enter the state name of U.S. which automatically marks itself on the map. it also returns a list of the states not guessed helping the user to correctly remember the states for the next time.